Abstract the response
